+++ src/simulation/allocation.jl
@@ -80,6 +80,7 @@ function allocate_arrays(ncells::Int, propellants, n_anom_vars::Int)
 
         # other caches
         cell_cache_1 = zeros(ncells),
+        cell_cache_2 = zeros(ncells),
 
         # Plume divergence variables
         channel_area = zeros(ncells),       # Area of channel / plume