+++ test/runtests.jl
@@ -2,7 +2,6 @@ using SafeTestsets
 using Documenter
 using HallThruster
 
-
 doctest(HallThruster)
 
 @safetestset "Thrusters" include("unit_tests/thrusters.jl")